DO NOT UPSTREAM: ARM64: Add cpu hotplug for ACPI parking method
This patch has been implemented for CPU hotplug support for ACPI parking
method.

Patch has been tested on mustang by offlining and then bringing back
online from sys interface. It has also been tested in kexec reboot.

It is just a workaround. Do not upstream it as it is, because it is not
following ACPI specs properly.

Signed-off-by: Pratyush Anand <panand@redhat.com>